Fail to write more than 50KB CSV.

+++ taiko/interface.py
@@ -13,6 +13,10 @@ class Interface(metaclass=_Singleton):
 
     def record_sensor(self, is_kill=False):
         if is_kill:
+            if self._pid is None:
+                print('no pid can be killed.')
+                return
+
             if platform.system() == 'Windows':
                 os.system('taskkill /F /PID ' + str(self._pid))
                 print('kill pid:%d ok' % self._pid)
@@ -20,6 +24,8 @@ class Interface(metaclass=_Singleton):
             elif platform.system() == 'Linux':
                 os.system('kill -9 ' + str(self._pid))
                 print('kill pid:%d ok' % self._pid)
+
+            self._pid = None
 
         else:
             capture_exe_path = posixpath.join(BASE_PATH, 'capture_sensor.py')
